feat: add forecast data to data layer

+++ app/src/main/java/com/immortalidiot/weathercompose/domain/model/DailyForecast.kt
@@ -2,10 +2,10 @@ package com.immortalidiot.weathercompose.domain.model
 
 data class DailyForecast(
     val date: String,
-    val averageTemp: Float,
+    val averageTemp: Int,
     val weatherDescription: String,
     val humidity: Int,
     val pressure: Int,
-    val windSpeed: Double,
+    val windSpeed: Int,
     val icon: String,
 )
